+++ src/Graph.hpp
@@ -16,14 +16,14 @@
 template <typename V>
 class Graph {
     protected:
-        typedef std::pair<unsigned int, std::vector<Node<V>*>> pair;
+        typedef std::pair<double, std::vector<Node<V>*>> pair;
         struct CustomCompare {
             bool operator()(const pair& first, const pair& second) {
                 return first.first > second.first;
             }
         };
         std::vector<Node<V>*> nodes;
-        std::vector<std::vector<unsigned int>> matrix;
+        std::vector<std::vector<double>> matrix;
         Node<V>* head = nullptr;
         std::map<Node<V>*, unsigned int> nodeMap;
 
@@ -42,18 +42,18 @@ class Graph {
 
 	// void printOut();
 	
-	void addNode(Node<V>* nextNode, std::vector<Node<V>*> adjacentNodes, unsigned int weight);
+	void addNode(Node<V>* nextNode, std::vector<Node<V>*> adjacentNodes, double weight);
 	void addNode(Node<V> *nextNode, std::vector<Node<V>*> adjacentNodes);
 	void addNode(Node<V> *nextNode, Node<V> *nodeITSLATE);
 
-	void addConnection(Node<V> *one, Node<V> *two, unsigned int weight);
+	void addConnection(Node<V> *one, Node<V> *two, double weight);
 	void addConnection(Node<V> *one, Node<V> *two);
-	void addNode(Node<V> *nextNode, Node<V> *nodeITSLATE, unsigned int weight); 
+	void addNode(Node<V> *nextNode, Node<V> *nodeITSLATE, double weight);
 	void setHead(int index);
 
         std::vector<Node<V>*> Dijkstra(Node<V>* from, Node<V>* find);
 
-	std::vector<std::vector<unsigned int>> getAdjacencyList();
+	std::vector<std::vector<double>> getAdjacencyList();
 
         std::vector<Node<V>*> getAdj(Node<V>* next);
